pyproject.toml
[project]
name = "ienf-q"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "huggingface-hub>=0.34.4",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "wandb>=0.21.3",
]
